Agrego implementacion de CMM

+++ src/winningPercentage.cpp
@@ -4,7 +4,7 @@ using namespace std;
 
 vector<double> winningPercentage::calculateLeaderboard(const int teamQuantity, const int matchesQuantity, const vector<vector<int>> resultMatrix) {
     vector<double> results(teamQuantity, 0.0);
-    vector<int> matchesQuantityPerTeam(teamQuantity, 0);
+    vector<double> matchesQuantityPerTeam(teamQuantity, 0);
 
     for (int i = 0; i < matchesQuantity; ++i) {
         matchesQuantityPerTeam[resultMatrix[i][1]-1]++;
